requirements.txt
absl-py==0.7.1
alabaster==0.7.12
allennlp==0.9.0
attrs==19.3.0
Babel==2.6.0
blis==0.2.4
boto3==1.9.122
botocore==1.12.122
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
Click==7.0
conllu==1.3.1
-e git+https://github.com/ConvLab/ConvLab.git@ea4d2a7bd10b60367ed760bdc21dcd4b087a54f3#egg=convlab
cuthon==0.5
cycler==0.10.0
cymem==2.0.3
Cython==0.29.10
docutils==0.14
editdistance==0.5.3
embeddings==0.0.6
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gevent==1.3.6
greenlet==0.4.15
h5py==2.10.0
icc-rt==2019.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
intel-openmp==2019.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
jsonnet==0.14.0
jsonpickle==1.2
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mkl==2019.0
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
murmurhash==1.0.2
nltk==3.4
numpy==1.17.2
numpydoc==0.8.0
olefile==0.46
overrides==2.6
packaging==19.0
pandas==0.25.3
parsimonious==0.8.0
Pillow==6.2.0
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
protobuf==3.11.1
py==1.8.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.3.1
pytest==5.3.2
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.1
regex==2019.12.19
requests==2.21.0
responses==0.10.6
s3transfer==0.2.0
sacremoses==0.0.35
scikit-learn==0.20.3
scipy==1.2.1
sentencepiece==0.1.85
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
spacy==2.1.9
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
srsly==0.2.0
tbb==2019.0
tbb4py==2019.0
tensorboardX==1.9
thinc==7.0.8
torch==1.3.1
torchvision==0.4.2
tqdm==4.31.1
transformers==2.2.2
Unidecode==1.1.1
urllib3==1.24.2
wasabi==0.4.2
wcwidth==0.1.7
Werkzeug==0.16.0
word2number==1.1
zipp==0.6.0
